requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pytesseract
pdf2image
pdfminer.six
google-generativeai